# ticketshow/api.yaml
openapi: 3.1.0
info:
  title: Ticket Show v2
  description: Modern Application development 2 project of roll number - 21f1006537
  version: "1"
servers:
  - url: http:/localhost:8080/
...
paths:
  /api/login:
    post:
      tags:
        - auth
      summary: Existing User logs In
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
              required:
                - email
                - password
      responses:
        "200":
          description: 
            User logged in successfully
  /api/register:
    post:
      tags:
        - auth
      summary: Registers New User
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                email:
                  type: string
                password:
                  type: string
              required:
                - username
                - email
                - password
      responses:
        "200":
          description: 
            New user created successfully
...